add compression/crc/encryption option on soeprotocol class

+++ src/crc.rs
@@ -3,8 +3,8 @@ use byteorder::{BigEndian, WriteBytesExt};
 use crc_table::get_crc_table;
 use wasm_bindgen::prelude::*;
 
-pub fn append_crc(data: &mut Vec<u8>, crc_seed: usize) -> () {
-    let crc = crc32(&data, crc_seed >> 0);
+pub fn append_crc(data: &mut Vec<u8>, crc_seed: u8) -> () {
+    let crc = crc32(&data, (crc_seed >> 0) as usize);
     data
         .write_u16::<BigEndian>((crc & 0xffff) as u16)
         .unwrap();
